pyproject.toml
[project]
name = "catalog"
version = "0.1.0"
description = "Catalog API"
authors = [
    {name = "Roman Bulgakov",email = "bulgakovra.w@gmail.com"}
]
readme = "README.md"
requires-python = "3.13.1"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dependency-injector (>=4.48.1,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pycasbin (>=2.2.0,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-factoryboy = "^2.8.1"
ruff = "^0.12.10"
pytest-mock = "^3.14.1"
pytest-asyncio = "^1.1.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
